# doc/openapi.yaml
openapi: 3.0.2
info:
 title: 'Auth di Express JS'
 version: 0.1.9
 description: 'Assignment week 11'
servers:
  - url: http://localhost:3000
    description: Development server
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: jwt
      description: 'Bearer with jwt  token'
  schemas:
...
    RegisterUser:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        role:
          type: string
    LoginUser:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    BadRequest:
      type: object
      properties:
        message:
          type: string
paths: 
...
  /register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUser'
      responses:
        '200':
          description: 'register user'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      password:
                        type: string
                        minLength: 8
                      role:
                        type: string
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
  /login:
    post:
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUser'
      responses:
        '200':
          description: 'login user'
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      username:
                        type: string
                      password:
                        type: string
                      role:
                        type: string
                      token:
                        type: string
        '400':
          description: 'Bad request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'  